nlp_project/solvers/gemini_client.py
import os

import google.generativeai as genai


if __name__ == "__main__":
    genai.configure(api_key=os.getenv('GOOGLE_API_KEY'))
    model = genai.GenerativeModel("gemini-1.5-flash")
    response = model.generate_content("Explain how AI works")
    print(response.text)

