pyproject.toml
[tool.poetry]
name = "nlp-final-project"
version = "0.1.0"
description = "NLP Final Project"
authors = [""]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pandas = "^2.2.3"
openai = "^1.60.1"
sympy = "^1.13.3"
google-generativeai = "^0.8.4"


[tool.poetry.dev-dependencies]
pytest = "^7.4.0"
black = "^23.9.0"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core>=1.5.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
